backend/requirements.txt
djangorestframework
djangorestframework-simplejwt
python-dotenv
